Rename loggingUtils into logging_utils
