+++ bin/active_learning_loop.py
@@ -27,7 +27,7 @@ from fink_utils.xmatch.simbad import return_list_of_eg_host
 
 from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession, load_parquet_files
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.classification import extract_fink_classification
 from fink_filters.filter_anomaly_notification.filter_utils import msg_handler_slack
