+++ bin/anomaly_archival.py
@@ -22,7 +22,7 @@ from fink_broker.sparkUtils import init_sparksession, load_parquet_files
 
 from fink_filters.filter_anomaly_notification.filter import anomaly_notification_
 
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_broker.hbase_utils import push_full_df_to_hbase, add_row_key
