+++ bin/daily_stats.py
@@ -25,7 +25,7 @@ import pyspark.sql.functions as F
 from fink_broker.sparkUtils import init_sparksession
 from fink_broker.hbase_utils import push_to_hbase
 from fink_broker.parser import getargs
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.classification import extract_fink_classification
 from fink_filters.filter_simbad_candidates.filter import simbad_candidates
