+++ bin/dwarf_agn.py
@@ -21,7 +21,7 @@ from pyspark.sql import functions as F
 
 from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession, load_parquet_files
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.filter_anomaly_notification.filter_utils import msg_handler_slack
 from fink_filters.filter_anomaly_notification.filter_utils import (
