+++ bin/index_sso_cand_archival.py
@@ -32,7 +32,7 @@ from fink_broker.sparkUtils import init_sparksession
 
 from fink_broker.hbase_utils import push_to_hbase
 
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 
 def main():
