+++ bin/save_distribution_schema.py
@@ -24,7 +24,7 @@ import json
 from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession, load_parquet_files
 from fink_broker.avro_utils import readschemafromavrofile
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 
 def main():
