+++ bin/science_archival.py
@@ -30,7 +30,7 @@ from fink_broker.sparkUtils import init_sparksession, load_parquet_files
 from fink_broker.hbase_utils import push_full_df_to_hbase
 from fink_broker.sparkUtils import list_hdfs_files
 
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 
 def main():
